requirements.txt
flask
sqlalchemy
flask-migrate
flask-admin
python-dotenv
mysql-connector
pycryptodome